backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20250625
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
regex==2025.7.34
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tiktoken==0.11.0
torch==2.2.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
